refactor: Renamed GenHashService into SecurityUtilityService
