+++ Services/SecurityUtility/Implementation/GeneratePassword.cs
@@ -1,8 +1,8 @@
 using System.Security.Cryptography;
 
-namespace FinBookeAPI.Services.GenHash;
+namespace FinBookeAPI.Services.SecurityUtility;
 
-public partial class GenHashService : IGenHashService
+public partial class SecurityUtilityService : ISecurityUtilityService
 {
     public string GeneratePassword(int length)
     {
