+++ Services/SecurityUtility/SecurityUtilityService.cs
@@ -1,11 +1,12 @@
 using System.Security.Cryptography;
 
-namespace FinBookeAPI.Services.GenHash;
+namespace FinBookeAPI.Services.SecurityUtility;
 
-public partial class GenHashService(ILogger<GenHashService> logger) : IGenHashService
+public partial class SecurityUtilityService(ILogger<SecurityUtilityService> logger)
+    : ISecurityUtilityService
 {
     private static readonly HashAlgorithmName _algorithm = HashAlgorithmName.SHA512;
-    private readonly ILogger<GenHashService> _logger = logger;
+    private readonly ILogger<SecurityUtilityService> _logger = logger;
     private const string _lowerCaseLetters = "abcdefghijklmnopqrstuvwxyz";
     private const string _upperCaseLetters = "ABCDEFGHIJKLMNOPQRSTUVWXYZ";
     private const string _digits = "0123456789";
