+++ Tests/SecurityUtility/HashUnitTest.cs
@@ -1,16 +1,16 @@
-using FinBookeAPI.Services.GenHash;
+using FinBookeAPI.Services.SecurityUtility;
 using Moq;
 
-namespace FinBookeAPI.Tests.GenHash;
+namespace FinBookeAPI.Tests.SecurityUtility;
 
 public partial class HashUnitTest
 {
-    private readonly GenHashService _service;
+    private readonly SecurityUtilityService _service;
 
     public HashUnitTest()
     {
-        var logger = new Mock<ILogger<GenHashService>>();
-        _service = new GenHashService(logger.Object);
+        var logger = new Mock<ILogger<SecurityUtilityService>>();
+        _service = new SecurityUtilityService(logger.Object);
     }
 
     [Fact]
